Refactor based on comments
