+++ VideoAPI/Video.API/Controllers/VideoEventsController.cs
@@ -5,12 +5,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Swashbuckle.AspNetCore.Annotations;
 using Video.API.Extensions;
+using Video.API.Mappings;
 using VideoApi.Contract.Requests;
-using VideoApi.DAL.Commands;
 using VideoApi.DAL.Commands.Core;
-using VideoApi.Domain.Enums;
 using VideoApi.Events.Handlers.Core;
-using VideoApi.Events.Models;
 
 namespace Video.API.Controllers
 {
@@ -46,7 +44,7 @@ namespace Video.API.Controllers
             Guid.TryParse(request.ConferenceId, out var conferenceId);
             Guid.TryParse(request.ParticipantId, out var participantId);
 
-            var command = MapEventRequestToEventCommand(conferenceId, request);
+            var command = EventRequestMapper.MapEventRequestToEventCommand(conferenceId, request);
 
             _logger.LogWarning("Handling {ConferenceEventRequest}", nameof(ConferenceEventRequest));
             
@@ -58,57 +56,9 @@ namespace Video.API.Controllers
                 return NoContent();
             }
 
-            var callbackEvent = MapEventRequestToEventHandlerDto(conferenceId, participantId, request);
+            var callbackEvent = EventRequestMapper.MapEventRequestToEventHandlerDto(conferenceId, participantId, request);
             await _eventHandlerFactory.Get(request.EventType).HandleAsync(callbackEvent);
             return NoContent();
-        }
-
-        private SaveEventCommand MapEventRequestToEventCommand(Guid conferenceId, ConferenceEventRequest request)
-        {
-            GetRoomTypeEnums(request, out var transferTo, out var transferFrom);
-
-            var command = new SaveEventCommand(conferenceId, request.EventId, request.EventType,
-                request.TimeStampUtc, transferFrom, transferTo, request.Reason, request.Phone);
-            if (Guid.TryParse(request.ParticipantId, out var participantId))
-            {
-                command.ParticipantId = participantId;
-            }
-
-            command.TransferredFromRoomLabel = request.TransferFrom;
-            command.TransferredToRoomLabel = request.TransferTo;
-
-            return command;
-        }
-
-        private CallbackEvent MapEventRequestToEventHandlerDto(Guid conferenceId, Guid participantId,
-            ConferenceEventRequest request)
-        {
-            GetRoomTypeEnums(request, out var transferTo, out var transferFrom);
-            
-            return new CallbackEvent
-            {
-                EventId = request.EventId,
-                EventType = request.EventType,
-                ConferenceId = conferenceId,
-                Reason = request.Reason,
-                TransferTo = transferTo,
-                TransferFrom = transferFrom,
-                TimeStampUtc = request.TimeStampUtc,
-                ParticipantId = participantId,
-                Phone = request.Phone,
-                TransferredFromRoomLabel = request.TransferFrom,
-                TransferredToRoomLabel = request.TransferTo
-            };
-        }
-
-        private static void GetRoomTypeEnums(ConferenceEventRequest request, out RoomType? transferTo,
-            out RoomType? transferFrom)
-        {
-            var isTransferFromEnum = Enum.TryParse(request.TransferFrom, out RoomType transferFromEnum);
-            var isTransferToEnum = Enum.TryParse(request.TransferTo, out RoomType transferToEnum);
-
-            transferFrom = isTransferFromEnum ? transferFromEnum : (RoomType?) null;
-            transferTo = isTransferToEnum ? transferToEnum : (RoomType?) null;
         }
     }
 }
