+++ VideoAPI/VideoApi.IntegrationTests/Database/Queries/GetAvailableRoomByRoomTypeQueryTests.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using VideoApi.DAL;
-using VideoApi.DAL.Commands;
 using VideoApi.DAL.Queries;
 using VideoApi.Domain;
 using VideoApi.Domain.Enums;
